% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppNode-class.R
\name{CppNode-class}
\alias{CppNode-class}
\alias{CppNode}
\alias{Rcpp_CppNode}
\alias{Rcpp_CppNode-class}
\title{C++ Quadtree Node}
\description{
The 'CppNode' is a C++ class represents a single quadtree node.
This is used internally - end users should have no need to use any of the
methods listed here.
}
\section{Fields}{

\describe{
\item{\code{asVector}}{\itemize{
  \item \strong{Description}: Returns a vector giving info about the node
  \item \strong{Parameters}: none
  \item \strong{Returns}: a numeric vector with the following named elements:
  \itemize{
    \item{\code{id}}
    \item{\code{hasChdn}}
    \item{\code{level}}
    \item{\code{xMin}}
    \item{\code{xMax}}
    \item{\code{yMin}}
    \item{\code{yMax}}
    \item{\code{smSide}}
  }
  \code{\link{as_data_frame}} makes use of this function to output info on
  each node - see the documentation of that function for details on the
  output
}}

\item{\code{getChildren}}{\itemize{
  \item \strong{Description}: Returns a \code{list} of the child nodes
  \item \strong{Parameters}: none
  \item \strong{Returns}: a \code{list} of \code{Rcpp_node} objects
}}

\item{\code{getNeighborIds}}{\itemize{
  \item \strong{Description}: Returns the IDs of the neighboring cells
  \item \strong{Parameters}: none
  \item \strong{Returns}: a numeric vector
}}

\item{\code{getNeighborInfo}}{\itemize{
  \item \strong{Description}: Returns a matrix with info on each of the
  neighboring cells
  \item \strong{Parameters}: none
  \item \strong{Returns}: a matrix. \code{quadtree$getNbList()} makes
  use of this function - see documentation of that function (by running
  \code{?'quadtree-class'}) for details on the return matrix.
}}

\item{\code{getNeighborVals}}{\itemize{
  \item \strong{Description}: Returns the values of all neighboring cells
  \item \strong{Parameters}: none
  \item \strong{Returns}: a numeric vector
}}

\item{\code{getNeighbors}}{\itemize{
  \item \strong{Description}: Returns a \code{list} of the neighboring nodes
  \item \strong{Parameters}: none
  \item \strong{Returns}: a \code{list} of \code{Rcpp_node} objects
}}

\item{\code{hasChildren}}{\itemize{
  \item \strong{Description}: Returns a boolean representing whether the node
  has children
  \item \strong{Parameters}: none
  \item \strong{Returns}: a boolean value - \code{TRUE} if it has children,
  \code{FALSE} otherwise
}}

\item{\code{id}}{\itemize{
  \item \strong{Description}: Returns the ID of this node
  \item \strong{Parameters}: none
  \item \strong{Returns}: an integer
}}

\item{\code{level}}{\itemize{
  \item \strong{Description}: Returns the 'level' (i.e. depth in the tree) of
  this node
  \item \strong{Parameters}: none
  \item \strong{Returns}: an integer
}}

\item{\code{smallestChildSideLength}}{\itemize{
  \item \strong{Description}: Returns the side length of the smallest
  descendant node
  \item \strong{Parameters}: none
  \item \strong{Returns}: a double
}}

\item{\code{value}}{\itemize{
  \item \strong{Description}: Returns the value of the node
  \item \strong{Parameters}: none
  \item \strong{Returns}: a double
}}

\item{\code{xLims}}{\itemize{
  \item \strong{Description}: Returns the x boundaries of the node
  \item \strong{Parameters}: none
  \item \strong{Returns}: 2-element numeric vector (xMin, xMax)
}}

\item{\code{yLims}}{\itemize{
  \item \strong{Description}: Returns the y boundaries of the node
  \item \strong{Parameters}: none
  \item \strong{Returns}: 2-element numeric vector (yMin, yMax)
}}
}}

