% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcp.R
\name{find_lcps}
\alias{find_lcps}
\alias{find_lcps,LcpFinder-method}
\title{Find LCPs to surrounding points}
\usage{
\S4method{find_lcps}{LcpFinder}(x, limit = NULL)
}
\arguments{
\item{x}{the \code{LcpFinder} object returned from \code{\link{lcp_finder}}}

\item{limit}{numeric; the maximum cost-distance for the LCPs. If \code{NULL},
no limit is applied and all possible LCPs (within the \code{LcpFinder}'s 
serach area) are found}
}
\value{
Returns a matrix summarizing each LCP found.
  \code{\link{summarize_lcps}} is used to generate this matrix - see the help
  for that function for details on the return matrix. Note that this function
  does \strong{not} return the full paths to each point - however, each of
  the paths summarized in the output matrix has already been calculated, and
  can be retrieved using \code{find_lcp()} (without having to recalculate
  the path, since it's already been calculated).
}
\description{
Calculates the LCPs to surrounding points. A constraint can be
  placed on the LCPs, so that only LCPs that are less than some specified
  cost-distance are returned.
}
\details{
A very important note to make is that once the LCP tree is calculated, it
never gets smaller. The implication of this is that great care is needed if
using a \code{\link{LcpFinder}} more than once. For example, I could use
\code{find_lcps(lcp_finder, limit=10)} to find all LCPs that have a
cost-distance less than 10. I could then use \code{summarize_lcps} to view
all cells that are reachable within 10 cost units. However, if I then run
\code{find_lcps(lcp_finder, limit=5)} to find all LCPs that have a
cost-distance less than 5, the underlying LCP network \strong{will remain
unchanged}. That is, if I run \code{summarize_lcps} on \code{lcp_finder}, it
will \strong{return paths with a cost-distance greater than 5}, since we had
previously used \code{lcp_finder} to find paths less than 10. As mentioned
before, this happens because the underlying data structure only ever adds
nodes, and never removes nodes.
}
\examples{
library(raster)

# create a quadtree
data(habitat)
rast <- habitat
qt <- quadtree(rast, split_threshold = .1, adj_type = "expand")

start_pt <- c(19000, 25000)

# finds LCPs to all cells
lcpf1 <- lcp_finder(qt, start_pt)
paths1 <- find_lcps(lcpf1, limit = NULL)

# limit LCPs by cost-distance
lcpf2 <- lcp_finder(qt, start_pt)
paths2 <- find_lcps(lcpf2, limit = 5000)

# Now plot the reachable cells
plot(qt, crop = TRUE, na_col = NULL, border_lwd = .3)
points(lcpf1, col = "black", pch = 16, cex = 1)
points(lcpf2, col = "red", pch = 16, cex = .7)
points(start_pt[1], start_pt[2], bg = "skyblue", col = "black", pch = 24,
       cex = 1.5)
}
\seealso{
\code{\link{lcp_finder}()} creates the \code{\link{LcpFinder}}
  object used as input to this function. \code{\link{find_lcp}()} returns the
  LCP between two points. \code{\link{summarize_lcps}()} outputs a summary
  matrix of all LCPs that have been calculated so far.
}
