% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{projection}
\alias{projection}
\alias{projection,Quadtree-method}
\alias{projection<-}
\alias{projection<-,Quadtree,ANY-method}
\alias{projection<-,Quadtree-method}
\title{Retrieve the proj4string of a quadtree}
\usage{
\S4method{projection}{Quadtree}(x)

\S4method{projection}{Quadtree}(x) <- value
}
\arguments{
\item{x}{a \code{\link{Quadtree}} object}

\item{value}{character; the projection to assign to the quadtree}
}
\value{
A character containing the proj4string
}
\description{
Retrieve the proj4string of a quadtree
}
