% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_as_data_frame.R
\name{qt_as_data_frame}
\alias{qt_as_data_frame}
\title{convert a quadtree to a data frame}
\usage{
qt_as_data_frame(quadtree)
}
\arguments{
\item{quadtree}{a quadtree object}
}
\value{
a data frame with one row for each quadtree cell. The columns are as
  follows:
  \itemize{
    \item \code{id}: the id of the cell
    \item \code{hasChdn}: 1 if the cell has children, 0 otherwise
    \item \code{level}: integer; the depth of this cell/node in the quadtree,
    where the root of the quadtree is considered to be level 0
    \item \code{xMin}, \code{xMax}, \code{yMin}, \code{yMax}: the x and y
    limits of the cell
    \item \code{value}: the value of the cell
    \item \code{smSide}: the smallest cell length among all of this cells
    descendants
    \item \code{parentID}: the ID of the cell's parent. The root, which has
    no parent, has a value of -1 for this element
  }
}
\description{
creates a data frame with information on each quadtree cell.
}
\examples{
mat = rbind(c(1,1,0,1),c(1,1,1,0),c(1,0,1,1),c(0,1,1,1))
qt = qt_create(mat,.1)
qt_plot(qt)
qt_as_data_frame(qt)
}
