% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_create.R
\name{qt_create}
\alias{qt_create}
\title{Create a quadtree from a raster}
\usage{
qt_create(rast, range_limit, adj_type = "expand", resample_n_side = NA)
}
\arguments{
\item{rast}{a \code{raster} object}

\item{range_limit}{numeric; if the cell values within a quadrant have a range larger than this value, the quadrant is split. See 'Details' for more}

\item{adj_type}{character; either \code{'expand'} or \code{'resample'}. See 'Details' for more.}

\item{resample_n_side}{integer; if \code{adj_type} is \code{'expand'}, this number is used to determine the dimensions to resample the raster to}
}
\description{
Create a quadtree from a raster
}
\details{
A quadtree is created from a raster by successively dividing the raster into
smaller and smaller cells, with the decision on whether to divide a cell
determined by \code{range_limit}. Initially, all of the cells in the raster
are considered. If the difference between the maximum and minimum cell values
exceeds \code{range_limit}, the raster is divided into four quadrants -
otherwise, the raster is not divided further and the mean of all values in
the raster is taken as the value for the resulting cell. Then, the process is
repeated for each of those 'child' cells, and then for their children, and so
on and so forth, until either \code{range_limit} is not exceeded or the
smallest possible cell size has been reached.

If a quadrant contains both NA cells and non-NA cells, that quadrant is
automatically divided. However, if a quadrant consists entirely of NA cells,
that cell is not divided further.

If a given quadrant has dimensions that are not divisible by 2 (for example,
5x5), then the process stops. Because of this, only rasters that have
dimensions that are a power of 2 can be divided down to their smallest cell
size. In addition, the rasters should be square.

To create quadtrees from rasters that have dimensions that are not a power of
two and are not square, two options are provided. The choice of method is
determined by the \code{adj_type} parameter.

In the 'expand' method, NA cells are added to the raster in order to create
an expanded raster whose dimensions are a power of 2. The smallest number
that is a power of two but greater than the larger dimension is used as the
dimensions of the expanded raster. For example, if a raster has dimensions
546 x 978, NA cells are added to the top and right of the raster in order to
create a raster with dimensions 1024 x 1024 (as 1024 is the smallest power of
2 that is also greater than 978).

In the 'resample' method, the raster is resampled in order to create a square
matrix with dimensions that are a power of two. There are two steps. First,
the raster must be made square. This is done in a way similar to the method
described above. The smaller dimension is padded with NA cells in order to
equal the larger dimension. For example, if the raster has dimensions 546 x
978, NA rows are added in order to create a raster with dimensions 978 x 978.
In the second step, this raster is then resampled to a user-specified
dimension (determined by the \code{resample_n_side} parameter). For example,
the user could set \code{resample_n_side} to be 1024, which will resample the
978 x 978 raster to 1024 x 1024. This raster can then be used to create a
quadtree.
}
\examples{
#create raster of random values
nrow = 57
ncol = 75
rast = raster(matrix(runif(nrow*ncol), nrow=nrow, ncol=ncol), xmn=0, xmx=ncol, ymn=0, ymx=nrow)

#create quadtree using the 'expand' method - automatically adds NA cells to
bring the dimensions to 128 x 128 before creating the quadtree
qt1 = qt_create(rast, range_limit = .9, adj_type="expand") 
qt_plot(qt1) #plot the quadtree
qt_plot(qt1, crop=TRUE) #we can use 'crop=TRUE' if we don't want to see the padded NA's

#create quadtree using the 'resample' method - we'll resample to 128 since it's a power of 2
qt2 = qt_create(rast, range_limit = .9, adj_type="resample", resample_n_side = 128)
qt_plot(qt2)
qt_plot(qt2, crop=TRUE)
}
