% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_lcp.R
\name{qt_find_lcp}
\alias{qt_find_lcp}
\title{Find the LCP between two points on a quadtree}
\usage{
qt_find_lcp(lcp_finder, end_point, use_original_end_points = FALSE)
}
\arguments{
\item{lcp_finder}{the LCP finder object returned from \code{qt_lcp_finder}}

\item{end_point}{numeric vector with two elements - the x and y coordinates
of the the destination point}

\item{use_original_end_points}{boolean; by default the start and end points
of the returned path are not the points given by the user but instead the
centroids of the cells that those points fall in. If this parameter is set
to \code{TRUE} the start and end points (representing the cell centroids)
are replaced with the actual points specified by the user. Note that this
is done after the calculation and has no effect on the path found by the
algorithm.}
}
\value{
\code{qt_find_lcp} returns a five column matrix representing the
  least cost path. It has the following columns: \itemize{ \item{\code{x}:
  }{x coordinate of this point} \item{\code{y}: }{y coordinate of this point}
  \item{\code{cost_tot}: }{the cumulative cost up to this point}
  \item{\code{dist_tot}: }{the cumulative distance up to this point - note
  that this is not straight-line distance, but instead the distance along the
  path} \item{\code{cost_cell}: }{the cost of the cell that contains this
  point} }

  If no path is possible between the two points, a 0-row matrix with the
  previously described columns is returned. Also, note that when creating the
  LCP finder object using \code{qt_lcp_finder}, \code{NULL} will be returned
  if \code{start_point} falls outside of the quadtree. If \code{NULL} is
  passed to the \code{lcp_finder} parameter, a 0-row matrix is returned.

  IMPORTANT NOTE: the \code{use_original_end_points} options ONLY changes the
  x and y coordinates of the first and last points - it doesn't change the
  \code{cost_tot} or \code{dist_tot} columns. This means that even though the
  start and end points have changed, the \code{cost_tot} and \code{dist_tot}
  columns still represent the cost and distance using the cell centroids of
  the start and end cells.
}
\description{
Finds the least cost path (LCP) between two points, using a
  quadtree as a resistance surface
}
\details{
See \code{\link{qt_lcp_finder}} for more information on how the LCP
  is found
}
\examples{
library(raster)

# ----- create a quadtree
# create raster of random values
nrow = 57
ncol = 75
set.seed(4)
rast = raster(matrix(runif(nrow*ncol), nrow=nrow, ncol=ncol), xmn=0, xmx=ncol, ymn=0, ymx=nrow)

# create quadtree
qt = qt_create(rast, range_limit = .9, adj_type="expand")

# defined our start and end points
start_pt = c(.231,.14)
end_pt = c(74.89,56.11)

# make an LCP finder object
lcpf = qt_lcp_finder(qt, start_pt)

# use the LCP finder object to find the LCP to a certain point
# this path will have the cell centroids as the start and end points
path1 = qt_find_lcp(lcpf, end_pt)
# this path will be identical to path1 except that the start and end points
# will be the user-provided start and end points rather than the cell centroids
path2 = qt_find_lcp(lcpf, end_pt, use_original_end_points = TRUE)

head(path1)
head(path2)

# plot the result
qt_plot(qt, crop=TRUE, border_col="gray60", na_col=NULL)
points(rbind(start_pt, end_pt), pch=16, col="red")
lines(path1[,1:2], col="black", lwd=2.5)
lines(path2[,1:2], col="red", lwd=1)
points(path1, cex=.7, pch=16)
}
\seealso{
\code{\link{qt_lcp_finder}}; \code{\link{qt_find_lcps}};
  \code{\link{qt_lcp_summary}}
}
