% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_lcp.R
\name{qt_lcp_summary}
\alias{qt_lcp_summary}
\title{Show a summary matrix of all LCPs currently calculated}
\usage{
qt_lcp_summary(lcp_finder)
}
\arguments{
\item{lcp_finder}{the LCP finder object returned from
\code{\link{qt_lcp_finder}}}
}
\value{
Returns a matrix
}
\description{
Given an LCP finder object, returns a matrix that summarizes all
  of the LCPs that have already been determined by the LCP finder.
}
