% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_proj4string.R
\name{qt_proj4string}
\alias{qt_proj4string}
\title{Retrieve the proj4string of a quadtree}
\usage{
qt_proj4string(quadtree)
}
\arguments{
\item{quadtree}{a quadtree object}
}
\value{
A character containing the proj4string
}
\description{
Retrieve the proj4string of a quadtree
}
