% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_read_write.R
\name{qt_read}
\alias{qt_read}
\alias{qt_write}
\title{Read/write a quadtree}
\usage{
qt_read(filepath)

qt_write(quadtree, filepath)
}
\arguments{
\item{filepath}{character; the filepath to read from or write to}

\item{quadtree}{quadtree object; the quadtree to write}
}
\description{
Read/write a quadtree
}
\details{
To read/write a quadtree object, the C++ library \code{cereal} is used to
serialize the quadtree and save it to a file. The file extension is
unimportant - it can be anything (I've been using the extension '.qtree').

Note that typically the quadtree isn't particularly space-efficient - it's
not uncommon for a quadtree file to be larger than the original raster file
(although, of course, this depends on how 'coarse' the quadtree is in
relation to the original raster). This is likely because the quadtree has to
store much more information about each cell (the x and y limits, its value,
pointers to its neighbors, among other things) while a raster can store only
the value since the coordinates of the cell can be determined from the
knowledge of the extent and the dimensions of the raster.

It's entirely possible that a quadtree implemention could be written that
is MUCH more space efficient. However, this was not the primary goal when 
creating this package.
}
\examples{
qt = qt_read("path/to/quadtree.qtree")
qt_write(qt, "path/to/newQuadtree.qtree")
}
