\name{quadtree-package}
\alias{quadtree-package}
\alias{quadtree}
\docType{package}
\title{
\packageTitle{quadtree}
}
\description{
\packageDescription{quadtree}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{quadtree}

Maintainer: \packageMaintainer{quadtree}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = quadtree)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{quadtree}}
%\section{Documentation Index}{\packageIndices{quadtree}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
