% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadtree-class.R
\name{shortestPathFinder-class}
\alias{shortestPathFinder-class}
\alias{Rcpp_shortestPathFinder}
\title{title here}
\description{
describe stuff
}
\section{Fields}{

\describe{
\item{\code{stuff}}{more stuff}
}}

