% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_Quadtree.R
\name{summary.Quadtree}
\alias{summary.Quadtree}
\alias{summary,Quadtree-method}
\alias{show.Quadtree}
\alias{show,Quadtree-method}
\title{Show a summary of a \code{\link{Quadtree}}}
\usage{
\S4method{summary}{Quadtree}(object)

\S4method{show}{Quadtree}(object)
}
\arguments{
\item{object}{a \code{\link{Quadtree}} object}
}
\value{
no return value
}
\description{
Prints out information about a \code{\link{Quadtree}}.
  Information shown is:
\itemize{
  \item class of object
  \item number of cells
  \item minimum cell size
  \item extent
  \item projection
  \item minimum and maximum values
}
}
\examples{
library(quadtree)
data(habitat)

qt <- quadtree(habitat, .1)
summary(qt)
}
