% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{summary}
\alias{summary}
\alias{summary,Quadtree-method}
\alias{show}
\alias{show,Quadtree-method}
\title{Show a Summary of a Quadtree Object}
\usage{
\S4method{summary}{Quadtree}(object)

\S4method{show}{Quadtree}(object)
}
\arguments{
\item{object}{a \code{\link{Quadtree}} object}
}
\value{
no return value
}
\description{
Prints out information about the quadtree. Information shown is:
\itemize{
  \item class of object
  \item number of cells
  \item minimum cell size
  \item extent
  \item projection
  \item minimum and maximum values
}
}
\examples{
data(habitat)
qt = quadtree(habitat,.1)
summary(qt)
}
