% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_foreign.R
\name{as_sf}
\alias{as_sf}
\alias{as_vect}
\alias{as_character}
\title{Convert to other R spatial objects}
\usage{
as_sf(x)

as_vect(x)

as_character(x)
}
\arguments{
\item{x}{Quadtree object}
}
\value{
an object of class `sf` or `SpatVector`, or a Well-Known Text (WKT) `character` representation
}
\description{
Convert to other R spatial objects
}
