% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_vector.R
\name{as_vector}
\alias{as_vector}
\alias{as_vector,Quadtree-method}
\title{Get All Cell Values in a Vector}
\usage{
\S4method{as_vector}{Quadtree}(x)
}
\arguments{
\item{x}{a \code{\link{Quadtree}} object}
}
\description{
Returns all cell values as a numeric vector
}
\examples{
data(habitat)
qt = quadtree(habitat,.2)
as_vector(qt)
}
\seealso{
\code{\link{as_data_frame}} creates a data frame from a quadtree
  that has all the cell values as well as details about each cell's size and
  extent.
}
