% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_extent.R
\name{qt_extent}
\alias{qt_extent}
\alias{qt_extent_orig}
\title{Get the extent of a quadtree}
\usage{
qt_extent(quadtree)

qt_extent_orig(quadtree)
}
\arguments{
\item{quadtree}{quadtree object}
}
\value{
Returns an 'extent' object
}
\description{
Gets the extent of the quadtree as an 'extent' object (from the 
raster package)
}
\details{
\code{qt_extent} returns the total extent covered by the quadtree.

\code{qt_extent_orig} returns the extent of the original raster used to
create the quadtree, before NA rows/columns were added to pad the dimensions.
This essentially represents the extent in which the non-NA data occurs.
}
\examples{
#create raster of random values
nrow = 57
ncol = 75
rast = raster(matrix(runif(nrow*ncol), nrow=nrow, ncol=ncol), xmn=0, xmx=ncol, ymn=0, ymx=nrow)
qt = qt_create(rast, .9, adj_type="expand")

qt_extent(qt)
qt_extent_orig(qt)
}
