% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Quadtree-class}
\alias{Quadtree-class}
\alias{Quadtree}
\title{Quadtree class}
\description{
This S4 class is a wrapper around a \code{CppQuadtree} C++ object that is
made available to R via the 'Rcpp' package. Instances of this class can be
created through the \code{\link{quadtree}()} function.

The methods of the C++ object (\code{\link{CppQuadtree}}) can be accessed
from R, but the typical end-user should have no need of these methods -
they are meant for internal use. That being said, descriptions of the
available methods can be found on the \code{\link{CppQuadtree}} documentation
page.
}
\section{Slots}{

\describe{
\item{\code{ptr}}{a C++ object of class \code{CppQuadtree}}
}}

