% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadtree-package.R
\docType{package}
\name{quadtree-package}
\alias{quadtree-package}
\title{Quadtree Representation of Rasters}
\description{
Provides a C++ implementation of a quadtree data
 structure. Functions are provided for creating a quadtree
 from a raster, with the level of "coarseness" of the
 quadtree being determined by a user-supplied parameter. In
 addition, functions are provided to extract values of the
 quadtree at point locations. Functions for calculating
 least-cost paths between two points are also provided.
}
\details{
To get an understanding of the most important aspects of the package, read
 the following:
   \enumerate{
     \item For details on how a quadtree is constructed, see
     \code{\link{quadtree}}.
     \item  For details on the least-cost path functionality, see
     \code{\link{lcp_finder}}.
   }
}
